Add template package in the permission.ws package
