+++ server/sonar-server/src/main/java/org/sonar/server/permission/ws/template/AddUserToTemplateAction.java
@@ -18,9 +18,10 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package org.sonar.server.permission.ws;
+package org.sonar.server.permission.ws.template;
 
-import javax.annotation.Nullable;
+import com.google.common.base.Predicate;
+import java.util.List;
 import org.sonar.api.server.ws.Request;
 import org.sonar.api.server.ws.Response;
 import org.sonar.api.server.ws.WebService;
@@ -28,26 +29,29 @@ import org.sonar.db.DbClient;
 import org.sonar.db.DbSession;
 import org.sonar.db.permission.PermissionQuery;
 import org.sonar.db.permission.PermissionTemplateDto;
-import org.sonar.db.user.GroupDto;
+import org.sonar.db.permission.UserWithPermissionDto;
+import org.sonar.db.user.UserDto;
+import org.sonar.server.permission.ws.PermissionDependenciesFinder;
+import org.sonar.server.permission.ws.PermissionsWsAction;
+import org.sonar.server.permission.ws.WsTemplateRef;
 import org.sonar.server.user.UserSession;
 
-import static org.sonar.api.security.DefaultGroups.ANYONE;
+import static com.google.common.collect.FluentIterable.from;
 import static org.sonar.db.user.GroupMembershipQuery.IN;
 import static org.sonar.server.permission.PermissionPrivilegeChecker.checkGlobalAdminUser;
 import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_PERMISSION;
-import static org.sonar.server.permission.ws.WsPermissionParameters.createGroupIdParameter;
-import static org.sonar.server.permission.ws.WsPermissionParameters.createGroupNameParameter;
+import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_USER_LOGIN;
 import static org.sonar.server.permission.ws.WsPermissionParameters.createProjectPermissionParameter;
 import static org.sonar.server.permission.ws.WsPermissionParameters.createTemplateParameters;
-import static org.sonar.server.permission.ws.PermissionRequestValidator.validateNotAnyoneAndAdminPermission;
+import static org.sonar.server.permission.ws.WsPermissionParameters.createUserLoginParameter;
 import static org.sonar.server.permission.ws.PermissionRequestValidator.validateProjectPermission;
 
-public class AddGroupToTemplateAction implements PermissionsWsAction {
+public class AddUserToTemplateAction implements PermissionsWsAction {
   private final DbClient dbClient;
   private final PermissionDependenciesFinder dependenciesFinder;
   private final UserSession userSession;
 
-  public AddGroupToTemplateAction(DbClient dbClient, PermissionDependenciesFinder dependenciesFinder, UserSession userSession) {
+  public AddUserToTemplateAction(DbClient dbClient, PermissionDependenciesFinder dependenciesFinder, UserSession userSession) {
     this.dbClient = dbClient;
     this.dependenciesFinder = dependenciesFinder;
     this.userSession = userSession;
@@ -56,18 +60,16 @@ public class AddGroupToTemplateAction implements PermissionsWsAction {
   @Override
   public void define(WebService.NewController context) {
     WebService.NewAction action = context
-      .createAction("add_group_to_template")
+      .createAction("add_user_to_template")
       .setPost(true)
       .setSince("5.2")
-      .setDescription("Add a group to a permission template.<br /> " +
-        "The group id or group name must be provided. <br />" +
+      .setDescription("Add a user to a permission template.<br /> " +
         "It requires administration permissions to access.")
       .setHandler(this);
 
     createTemplateParameters(action);
     createProjectPermissionParameter(action);
-    createGroupIdParameter(action);
-    createGroupNameParameter(action);
+    createUserLoginParameter(action);
   }
 
   @Override
@@ -75,19 +77,16 @@ public class AddGroupToTemplateAction implements PermissionsWsAction {
     checkGlobalAdminUser(userSession);
 
     String permission = wsRequest.mandatoryParam(PARAM_PERMISSION);
-    WsGroupRef group = WsGroupRef.fromRequest(wsRequest);
+    final String userLogin = wsRequest.mandatoryParam(PARAM_USER_LOGIN);
 
     DbSession dbSession = dbClient.openSession(false);
     try {
       validateProjectPermission(permission);
-      validateNotAnyoneAndAdminPermission(permission, group.name());
-
       PermissionTemplateDto template = dependenciesFinder.getTemplate(dbSession, WsTemplateRef.fromRequest(wsRequest));
-      GroupDto groupDto = dependenciesFinder.getGroup(dbSession, group);
+      UserDto user = dependenciesFinder.getUser(dbSession, userLogin);
 
-      if (!groupAlreadyAdded(dbSession, template.getId(), groupDto, permission)) {
-        Long groupId = groupDto == null ? null : groupDto.getId();
-        dbClient.permissionTemplateDao().insertGroupPermission(dbSession, template.getId(), groupId, permission);
+      if (!isUserAlreadyAdded(dbSession, template.getId(), userLogin, permission)) {
+        dbClient.permissionTemplateDao().insertUserPermission(dbSession, template.getId(), user.getId(), permission);
       }
     } finally {
       dbClient.closeSession(dbSession);
@@ -96,9 +95,22 @@ public class AddGroupToTemplateAction implements PermissionsWsAction {
     wsResponse.noContent();
   }
 
-  private boolean groupAlreadyAdded(DbSession dbSession, long templateId, @Nullable GroupDto group, String permission) {
-    String groupName = group == null ? ANYONE : group.getName();
-    PermissionQuery permissionQuery = PermissionQuery.builder().membership(IN).permission(permission).build();
-    return dbClient.permissionTemplateDao().hasGroup(dbSession, permissionQuery, templateId, groupName);
+  private boolean isUserAlreadyAdded(DbSession dbSession, long templateId, String userLogin, String permission) {
+    PermissionQuery permissionQuery = PermissionQuery.builder().permission(permission).membership(IN).build();
+    List<UserWithPermissionDto> usersWithPermission = dbClient.permissionTemplateDao().selectUsers(dbSession, permissionQuery, templateId, 0, Integer.MAX_VALUE);
+    return from(usersWithPermission).anyMatch(new HasUserPredicate(userLogin));
+  }
+
+  private static class HasUserPredicate implements Predicate<UserWithPermissionDto> {
+    private final String userLogin;
+
+    public HasUserPredicate(String userLogin) {
+      this.userLogin = userLogin;
+    }
+
+    @Override
+    public boolean apply(UserWithPermissionDto userWithPermission) {
+      return userLogin.equals(userWithPermission.getLogin());
+    }
   }
 }
