+++ server/sonar-server/src/main/resources/org/sonar/server/permission/ws/template/create_template-example.json
