+++ server/sonar-server/src/main/resources/org/sonar/server/permission/ws/template/search_templates-example.json
