+++ server/sonar-server/src/main/resources/org/sonar/server/permission/ws/template/update_template-example.json
