+++ server/sonar-server/src/test/java/org/sonar/server/permission/ws/template/CreateTemplateActionTest.java
@@ -18,9 +18,8 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package org.sonar.server.permission.ws;
+package org.sonar.server.permission.ws.template;
 
-import java.util.Date;
 import javax.annotation.Nullable;
 import org.junit.Before;
 import org.junit.Rule;
@@ -32,11 +31,8 @@ import org.sonar.db.DbClient;
 import org.sonar.db.DbSession;
 import org.sonar.db.DbTester;
 import org.sonar.db.permission.PermissionTemplateDto;
-import org.sonar.db.user.GroupDto;
-import org.sonar.server.component.ComponentFinder;
 import org.sonar.server.exceptions.BadRequestException;
 import org.sonar.server.exceptions.ForbiddenException;
-import org.sonar.server.exceptions.NotFoundException;
 import org.sonar.server.exceptions.UnauthorizedException;
 import org.sonar.server.tester.UserSessionRule;
 import org.sonar.server.ws.TestRequest;
@@ -48,12 +44,11 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.sonar.db.permission.PermissionTemplateTesting.newPermissionTemplateDto;
 import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_DESCRIPTION;
-import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_ID;
 import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_NAME;
 import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_PATTERN;
 import static org.sonar.test.JsonAssert.assertJson;
 
-public class UpdateTemplateActionTest {
+public class CreateTemplateActionTest {
 
   @Rule
   public DbTester db = DbTester.create(System2.INSTANCE);
@@ -65,94 +60,52 @@ public class UpdateTemplateActionTest {
   WsActionTester ws;
   DbClient dbClient;
   DbSession dbSession;
-
-  PermissionTemplateDto templateDto;
+  System2 system = mock(System2.class);
 
   @Before
   public void setUp() {
     userSession.login().setGlobalPermissions(GlobalPermissions.SYSTEM_ADMIN);
-    System2 system = mock(System2.class);
-    when(system.now()).thenReturn(1_440_512_328_743L);
+    when(system.now()).thenReturn(1440512328743L);
 
     dbClient = db.getDbClient();
     dbSession = db.getSession();
-    PermissionDependenciesFinder finder = new PermissionDependenciesFinder(dbClient, new ComponentFinder(dbClient));
-
-    ws = new WsActionTester(new UpdateTemplateAction(dbClient, userSession, system, finder));
-
-    templateDto = insertTemplate(newPermissionTemplateDto()
-      .setName("Permission Template Name")
-      .setDescription("Permission Template Description")
-      .setKeyPattern(".*\\.pattern\\..*")
-      .setCreatedAt(new Date(1_000_000_000_000L))
-      .setUpdatedAt(new Date(1_000_000_000_000L)));
-    commit();
+    ws = new WsActionTester(new CreateTemplateAction(dbClient, userSession, system));
   }
 
   @Test
-  public void update_all_permission_template_fields() {
-    TestResponse result = newRequest(templateDto.getKee(), "Finance", "Permissions for financially related projects", ".*\\.finance\\..*");
+  public void create_full_permission_template() {
+    TestResponse result = newRequest("Finance", "Permissions for financially related projects", ".*\\.finance\\..*");
 
     assertJson(result.getInput())
       .ignoreFields("id")
-      .isSimilarTo(getClass().getResource("update_template-example.json"));
+      .isSimilarTo(getClass().getResource("create_template-example.json"));
     PermissionTemplateDto finance = dbClient.permissionTemplateDao().selectByName(dbSession, "Finance");
     assertThat(finance.getName()).isEqualTo("Finance");
     assertThat(finance.getDescription()).isEqualTo("Permissions for financially related projects");
     assertThat(finance.getKeyPattern()).isEqualTo(".*\\.finance\\..*");
-    assertThat(finance.getKee()).isEqualTo(templateDto.getKee());
-    assertThat(finance.getCreatedAt()).isEqualTo(templateDto.getCreatedAt());
+    assertThat(finance.getKee()).isNotEmpty();
+    assertThat(finance.getCreatedAt().getTime()).isEqualTo(1440512328743L);
     assertThat(finance.getUpdatedAt().getTime()).isEqualTo(1440512328743L);
   }
 
   @Test
-  public void update_with_the_same_values() {
-    newRequest(templateDto.getKee(), templateDto.getName(), templateDto.getDescription(), templateDto.getKeyPattern());
-
-    PermissionTemplateDto updatedTemplate = dbClient.permissionTemplateDao().selectByUuid(dbSession, templateDto.getKee());
-    assertThat(updatedTemplate.getName()).isEqualTo(templateDto.getName());
-    assertThat(updatedTemplate.getDescription()).isEqualTo(templateDto.getDescription());
-    assertThat(updatedTemplate.getKeyPattern()).isEqualTo(templateDto.getKeyPattern());
-  }
-
-  @Test
-  public void update_name_only() {
-    newRequest(templateDto.getKee(), "Finance", null, null);
+  public void create_minimalist_permission_template() {
+    newRequest("Finance", null, null);
 
     PermissionTemplateDto finance = dbClient.permissionTemplateDao().selectByName(dbSession, "Finance");
     assertThat(finance.getName()).isEqualTo("Finance");
-    assertThat(finance.getDescription()).isEqualTo(templateDto.getDescription());
-    assertThat(finance.getKeyPattern()).isEqualTo(templateDto.getKeyPattern());
+    assertThat(finance.getDescription()).isNullOrEmpty();
+    assertThat(finance.getKeyPattern()).isNullOrEmpty();
+    assertThat(finance.getKee()).isNotEmpty();
+    assertThat(finance.getCreatedAt().getTime()).isEqualTo(1440512328743L);
+    assertThat(finance.getUpdatedAt().getTime()).isEqualTo(1440512328743L);
   }
 
   @Test
-  public void fail_if_key_is_not_found() {
-    expectedException.expect(NotFoundException.class);
-    expectedException.expectMessage("Permission template with id 'unknown-key' is not found");
-
-    newRequest("unknown-key", null, null, null);
-  }
-
-  @Test
-  public void fail_if_name_already_exists_in_another_template() {
-    expectedException.expect(BadRequestException.class);
-    expectedException.expectMessage("A template with the name 'My Template' already exists (case insensitive).");
-
-    insertTemplate(newPermissionTemplateDto()
-      .setName("My Template")
-      .setKee("my-key")
-      .setCreatedAt(new Date(12345789L))
-      .setUpdatedAt(new Date(12345789L)));
-    commit();
-
-    newRequest(templateDto.getKee(), "My Template", null, null);
-  }
-
-  @Test
-  public void fail_if_key_is_not_provided() {
+  public void fail_if_name_not_provided() {
     expectedException.expect(IllegalArgumentException.class);
 
-    newRequest(null, "Finance", null, null);
+    newRequest(null, null, null);
   }
 
   @Test
@@ -160,15 +113,7 @@ public class UpdateTemplateActionTest {
     expectedException.expect(BadRequestException.class);
     expectedException.expectMessage("The template name must not be blank");
 
-    newRequest(templateDto.getKee(), "", null, null);
-  }
-
-  @Test
-  public void fail_if_name_has_just_whitespaces() {
-    expectedException.expect(BadRequestException.class);
-    expectedException.expectMessage("The template name must not be blank");
-
-    newRequest(templateDto.getKee(), "  \r\n", null, null);
+    newRequest("", null, null);
   }
 
   @Test
@@ -176,7 +121,7 @@ public class UpdateTemplateActionTest {
     expectedException.expect(BadRequestException.class);
     expectedException.expectMessage("The 'projectKeyPattern' parameter must be a valid Java regular expression. '[azerty' was passed");
 
-    newRequest(templateDto.getKee(), "Finance", null, "[azerty");
+    newRequest("Finance", null, "[azerty");
   }
 
   @Test
@@ -186,7 +131,7 @@ public class UpdateTemplateActionTest {
     insertTemplate(newPermissionTemplateDto().setName("finance"));
     commit();
 
-    newRequest(templateDto.getKee(), "Finance", null, null);
+    newRequest("Finance", null, null);
   }
 
   @Test
@@ -194,7 +139,7 @@ public class UpdateTemplateActionTest {
     expectedException.expect(UnauthorizedException.class);
     userSession.anonymous();
 
-    newRequest(templateDto.getKee(), "Finance", null, null);
+    newRequest("Finance", null, null);
   }
 
   @Test
@@ -202,26 +147,19 @@ public class UpdateTemplateActionTest {
     expectedException.expect(ForbiddenException.class);
     userSession.setGlobalPermissions(GlobalPermissions.QUALITY_PROFILE_ADMIN);
 
-    newRequest(templateDto.getKee(), "Finance", null, null);
+    newRequest("Finance", null, null);
   }
 
   private PermissionTemplateDto insertTemplate(PermissionTemplateDto template) {
     return dbClient.permissionTemplateDao().insert(dbSession, template);
   }
 
-  private GroupDto insertGroup(GroupDto group) {
-    return dbClient.groupDao().insert(db.getSession(), group);
-  }
-
   private void commit() {
     dbSession.commit();
   }
 
-  private TestResponse newRequest(@Nullable String key, @Nullable String name, @Nullable String description, @Nullable String projectPattern) {
+  private TestResponse newRequest(@Nullable String name, @Nullable String description, @Nullable String projectPattern) {
     TestRequest request = ws.newRequest();
-    if (key != null) {
-      request.setParam(PARAM_ID, key);
-    }
     if (name != null) {
       request.setParam(PARAM_NAME, name);
     }
