+++ server/sonar-server/src/test/java/org/sonar/server/permission/ws/template/UpdateTemplateActionTest.java
@@ -18,8 +18,9 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package org.sonar.server.permission.ws;
+package org.sonar.server.permission.ws.template;
 
+import java.util.Date;
 import javax.annotation.Nullable;
 import org.junit.Before;
 import org.junit.Rule;
@@ -31,9 +32,13 @@ import org.sonar.db.DbClient;
 import org.sonar.db.DbSession;
 import org.sonar.db.DbTester;
 import org.sonar.db.permission.PermissionTemplateDto;
+import org.sonar.db.user.GroupDto;
+import org.sonar.server.component.ComponentFinder;
 import org.sonar.server.exceptions.BadRequestException;
 import org.sonar.server.exceptions.ForbiddenException;
+import org.sonar.server.exceptions.NotFoundException;
 import org.sonar.server.exceptions.UnauthorizedException;
+import org.sonar.server.permission.ws.PermissionDependenciesFinder;
 import org.sonar.server.tester.UserSessionRule;
 import org.sonar.server.ws.TestRequest;
 import org.sonar.server.ws.TestResponse;
@@ -44,11 +49,12 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.sonar.db.permission.PermissionTemplateTesting.newPermissionTemplateDto;
 import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_DESCRIPTION;
+import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_ID;
 import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_NAME;
 import static org.sonar.server.permission.ws.WsPermissionParameters.PARAM_PATTERN;
 import static org.sonar.test.JsonAssert.assertJson;
 
-public class CreateTemplateActionTest {
+public class UpdateTemplateActionTest {
 
   @Rule
   public DbTester db = DbTester.create(System2.INSTANCE);
@@ -60,52 +66,94 @@ public class CreateTemplateActionTest {
   WsActionTester ws;
   DbClient dbClient;
   DbSession dbSession;
-  System2 system = mock(System2.class);
+
+  PermissionTemplateDto templateDto;
 
   @Before
   public void setUp() {
     userSession.login().setGlobalPermissions(GlobalPermissions.SYSTEM_ADMIN);
-    when(system.now()).thenReturn(1440512328743L);
+    System2 system = mock(System2.class);
+    when(system.now()).thenReturn(1_440_512_328_743L);
 
     dbClient = db.getDbClient();
     dbSession = db.getSession();
-    ws = new WsActionTester(new CreateTemplateAction(dbClient, userSession, system));
+    PermissionDependenciesFinder finder = new PermissionDependenciesFinder(dbClient, new ComponentFinder(dbClient));
+
+    ws = new WsActionTester(new UpdateTemplateAction(dbClient, userSession, system, finder));
+
+    templateDto = insertTemplate(newPermissionTemplateDto()
+      .setName("Permission Template Name")
+      .setDescription("Permission Template Description")
+      .setKeyPattern(".*\\.pattern\\..*")
+      .setCreatedAt(new Date(1_000_000_000_000L))
+      .setUpdatedAt(new Date(1_000_000_000_000L)));
+    commit();
   }
 
   @Test
-  public void create_full_permission_template() {
-    TestResponse result = newRequest("Finance", "Permissions for financially related projects", ".*\\.finance\\..*");
+  public void update_all_permission_template_fields() {
+    TestResponse result = newRequest(templateDto.getKee(), "Finance", "Permissions for financially related projects", ".*\\.finance\\..*");
 
     assertJson(result.getInput())
       .ignoreFields("id")
-      .isSimilarTo(getClass().getResource("create_template-example.json"));
+      .isSimilarTo(getClass().getResource("update_template-example.json"));
     PermissionTemplateDto finance = dbClient.permissionTemplateDao().selectByName(dbSession, "Finance");
     assertThat(finance.getName()).isEqualTo("Finance");
     assertThat(finance.getDescription()).isEqualTo("Permissions for financially related projects");
     assertThat(finance.getKeyPattern()).isEqualTo(".*\\.finance\\..*");
-    assertThat(finance.getKee()).isNotEmpty();
-    assertThat(finance.getCreatedAt().getTime()).isEqualTo(1440512328743L);
+    assertThat(finance.getKee()).isEqualTo(templateDto.getKee());
+    assertThat(finance.getCreatedAt()).isEqualTo(templateDto.getCreatedAt());
     assertThat(finance.getUpdatedAt().getTime()).isEqualTo(1440512328743L);
   }
 
   @Test
-  public void create_minimalist_permission_template() {
-    newRequest("Finance", null, null);
+  public void update_with_the_same_values() {
+    newRequest(templateDto.getKee(), templateDto.getName(), templateDto.getDescription(), templateDto.getKeyPattern());
+
+    PermissionTemplateDto updatedTemplate = dbClient.permissionTemplateDao().selectByUuid(dbSession, templateDto.getKee());
+    assertThat(updatedTemplate.getName()).isEqualTo(templateDto.getName());
+    assertThat(updatedTemplate.getDescription()).isEqualTo(templateDto.getDescription());
+    assertThat(updatedTemplate.getKeyPattern()).isEqualTo(templateDto.getKeyPattern());
+  }
+
+  @Test
+  public void update_name_only() {
+    newRequest(templateDto.getKee(), "Finance", null, null);
 
     PermissionTemplateDto finance = dbClient.permissionTemplateDao().selectByName(dbSession, "Finance");
     assertThat(finance.getName()).isEqualTo("Finance");
-    assertThat(finance.getDescription()).isNullOrEmpty();
-    assertThat(finance.getKeyPattern()).isNullOrEmpty();
-    assertThat(finance.getKee()).isNotEmpty();
-    assertThat(finance.getCreatedAt().getTime()).isEqualTo(1440512328743L);
-    assertThat(finance.getUpdatedAt().getTime()).isEqualTo(1440512328743L);
+    assertThat(finance.getDescription()).isEqualTo(templateDto.getDescription());
+    assertThat(finance.getKeyPattern()).isEqualTo(templateDto.getKeyPattern());
   }
 
   @Test
-  public void fail_if_name_not_provided() {
+  public void fail_if_key_is_not_found() {
+    expectedException.expect(NotFoundException.class);
+    expectedException.expectMessage("Permission template with id 'unknown-key' is not found");
+
+    newRequest("unknown-key", null, null, null);
+  }
+
+  @Test
+  public void fail_if_name_already_exists_in_another_template() {
+    expectedException.expect(BadRequestException.class);
+    expectedException.expectMessage("A template with the name 'My Template' already exists (case insensitive).");
+
+    insertTemplate(newPermissionTemplateDto()
+      .setName("My Template")
+      .setKee("my-key")
+      .setCreatedAt(new Date(12345789L))
+      .setUpdatedAt(new Date(12345789L)));
+    commit();
+
+    newRequest(templateDto.getKee(), "My Template", null, null);
+  }
+
+  @Test
+  public void fail_if_key_is_not_provided() {
     expectedException.expect(IllegalArgumentException.class);
 
-    newRequest(null, null, null);
+    newRequest(null, "Finance", null, null);
   }
 
   @Test
@@ -113,7 +161,15 @@ public class CreateTemplateActionTest {
     expectedException.expect(BadRequestException.class);
     expectedException.expectMessage("The template name must not be blank");
 
-    newRequest("", null, null);
+    newRequest(templateDto.getKee(), "", null, null);
+  }
+
+  @Test
+  public void fail_if_name_has_just_whitespaces() {
+    expectedException.expect(BadRequestException.class);
+    expectedException.expectMessage("The template name must not be blank");
+
+    newRequest(templateDto.getKee(), "  \r\n", null, null);
   }
 
   @Test
@@ -121,7 +177,7 @@ public class CreateTemplateActionTest {
     expectedException.expect(BadRequestException.class);
     expectedException.expectMessage("The 'projectKeyPattern' parameter must be a valid Java regular expression. '[azerty' was passed");
 
-    newRequest("Finance", null, "[azerty");
+    newRequest(templateDto.getKee(), "Finance", null, "[azerty");
   }
 
   @Test
@@ -131,7 +187,7 @@ public class CreateTemplateActionTest {
     insertTemplate(newPermissionTemplateDto().setName("finance"));
     commit();
 
-    newRequest("Finance", null, null);
+    newRequest(templateDto.getKee(), "Finance", null, null);
   }
 
   @Test
@@ -139,7 +195,7 @@ public class CreateTemplateActionTest {
     expectedException.expect(UnauthorizedException.class);
     userSession.anonymous();
 
-    newRequest("Finance", null, null);
+    newRequest(templateDto.getKee(), "Finance", null, null);
   }
 
   @Test
@@ -147,19 +203,26 @@ public class CreateTemplateActionTest {
     expectedException.expect(ForbiddenException.class);
     userSession.setGlobalPermissions(GlobalPermissions.QUALITY_PROFILE_ADMIN);
 
-    newRequest("Finance", null, null);
+    newRequest(templateDto.getKee(), "Finance", null, null);
   }
 
   private PermissionTemplateDto insertTemplate(PermissionTemplateDto template) {
     return dbClient.permissionTemplateDao().insert(dbSession, template);
   }
 
+  private GroupDto insertGroup(GroupDto group) {
+    return dbClient.groupDao().insert(db.getSession(), group);
+  }
+
   private void commit() {
     dbSession.commit();
   }
 
-  private TestResponse newRequest(@Nullable String name, @Nullable String description, @Nullable String projectPattern) {
+  private TestResponse newRequest(@Nullable String key, @Nullable String name, @Nullable String description, @Nullable String projectPattern) {
     TestRequest request = ws.newRequest();
+    if (key != null) {
+      request.setParam(PARAM_ID, key);
+    }
     if (name != null) {
       request.setParam(PARAM_NAME, name);
     }
